pyproject.toml
[tool.poetry]
name = "fhr-deal-finder"
version = "0.1.0"
description = ""
authors = ["Tristan <tkw6eh@virginia.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
certifi = "2024.6.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
cloudscraper = "1.2.71"
idna = "3.7"
numpy = "2.0.0"
packaging = "24.1"
pandas = "2.2.2"
plotly = "5.22.0"
pyparsing = "3.1.2"
python-dateutil = "2.9.0.post0"
pytz = "2024.1"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
six = "1.16.0"
tenacity = "8.4.1"
tzdata = "2024.1"
urllib3 = "2.2.2"
prompt-toolkit = "^3.0.48"
pytest-playwright = "^0.5.2"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
